softmaxxing all models

+++ imagedetect/trainer.py
@@ -135,7 +135,7 @@ class Trainer:
         predicted = T.argmax(output, dim=1)
         # print(predicted, target)
         # compare with the target tensor to get a tensor of correct predictions
-        correct = (predicted == target.to(correct.dtype))
+        correct = (predicted == target.to(predicted.dtype))
         # calculate the accuracy as the percentage of correct predictions
         accuracy = correct.float().mean()
         return accuracy
